Remove Unbound, enable Split Tunnel, local PiHole
Running this on a Raspberry Pi 2, I noticed that Unbound was noticeably
slow. Furthermore, you had to adjust the image for the service to a
Raspberry Pi-specific one.

I opted to just remove Unbound as the benefits it provided were not
worth it to me compared to the downsides in my setup.

Additionally, this enables a Split Tunnel by default and exposes the
ports for PiHole to the default ones so that it can operate for clients
on your Local Network without necessitating WireGuard.